small code style fixes
