+++ .flake8
@@ -7,3 +7,4 @@ select = B,C,E,F,W,T4,B9,Q0,N8,VNE
 exclude =
     **migrations
     .venv
+    tests
