+++ templates/base.html
@@ -5,6 +5,7 @@
   {% block title %}<title>Taxi Service</title>{% endblock %}
   <meta charset="utf-8">
   <meta name="viewport" content="width=device-width, initial-scale=1">
+  <title>Hotel reviews</title>
   {% load static %}
   <link rel="stylesheet" href="{% static 'css/styles.css' %}">
   <link rel="stylesheet" href="{% static 'css/bootstrap.css' %}">
